todo_react/requirements.txt
Django==2.2
django-cors-headers==3.6.0
djangorestframework==3.12.2
pytz==2020.5
sqlparse==0.4.1
